Move i18n code to subfolder
