+++ .tx/config
@@ -2,8 +2,8 @@
 host = https://www.transifex.com
 
 [o:joinmarket:p:jam:r:translation]
-file_filter  = src/locales/<lang>/translation.json
-source_file  = src/locales/en/translation.json
+file_filter  = src/i18n/locales/<lang>/translation.json
+source_file  = src/i18n/locales/en/translation.json
 type         = KEYVALUEJSON
 minimum_perc = 0
