+++ src/i18n/test_config.js
@@ -1,0 +1,10 @@
+import i18n from 'i18next'
+import { initReactI18next } from 'react-i18next'
+
+i18n.use(initReactI18next).init({
+  lng: 'en',
+  fallbackLng: 'en',
+  resources: { en: { translations: {} } },
+})
+
+export default i18n
